move Network struct inside teh dashcore crate
