--- dash-network-ffi/Cargo.toml
@@ -1,21 +1,0 @@
-[package]
-name = "dash-network-ffi"
-version = { workspace = true }
-edition = "2021"
-authors = ["Quantum Explorer <quantum@dash.org>"]
-license = "CC0-1.0"
-repository = "https://github.com/dashpay/rust-dashcore/"
-description = "FFI bindings for dash-network types"
-keywords = ["dash", "network", "ffi", "bindings"]
-readme = "README.md"
-
-[dependencies]
-dash-network = { path = "../dash-network", default-features = false }
-thiserror = "2.0.12"
-
-[dev-dependencies]
-hex = "0.4"
-
-[lib]
-crate-type = ["cdylib", "staticlib"]
-name = "dash_network_ffi"
