--- dash-network-ffi/build.rs
@@ -1,4 +1,0 @@
-fn main() {
-    // Build script for dash-network-ffi
-    // Standard FFI compilation without uniffi
-}
