--- dash-network-ffi/src/dash_network_ffiFFI.modulemap
@@ -1,7 +1,0 @@
-module dash_network_ffiFFI {
-    header "dash_network_ffiFFI.h"
-    export *
-    use "Darwin"
-    use "_Builtin_stdbool"
-    use "_Builtin_stdint"
-}
