+++ dash/Cargo.toml
@@ -23,7 +23,7 @@ default = [ "std", "secp-recovery", "bincode" ]
 base64 = [ "base64-compat" ]
 rand-std = ["secp256k1/rand"]
 rand = ["secp256k1/rand"]
-serde = ["dep:serde", "dashcore_hashes/serde", "secp256k1/serde", "dash-network/serde"]
+serde = ["dep:serde", "dashcore_hashes/serde", "secp256k1/serde"]
 secp-lowmemory = ["secp256k1/lowmemory"]
 secp-recovery = ["secp256k1/recovery"]
 signer = ["secp-recovery", "rand", "base64"]
@@ -32,7 +32,7 @@ bls = ["blsful"]
 eddsa = ["ed25519-dalek"]
 quorum_validation = ["bls"]
 message_verification = ["bls"]
-bincode = [ "dep:bincode", "dep:bincode_derive", "dashcore_hashes/bincode", "dash-network/bincode" ]
+bincode = [ "dep:bincode", "dep:bincode_derive", "dashcore_hashes/bincode" ]
 test-utils = []
 
 # At least one of std, no-std must be enabled.
@@ -40,8 +40,8 @@ test-utils = []
 # The no-std feature doesn't disable std - you need to turn off the std feature for that by disabling default.
 # Instead no-std enables additional features required for this crate to be usable without std.
 # As a result, both can be enabled without conflict.
-std = ["secp256k1/std", "dashcore_hashes/std", "bech32/std", "internals/std", "dash-network/std"]
-no-std = ["core2", "dashcore_hashes/alloc", "dashcore_hashes/core2", "secp256k1/alloc", "dash-network/no-std"]
+std = ["secp256k1/std", "dashcore_hashes/std", "bech32/std", "internals/std"]
+no-std = ["core2", "dashcore_hashes/alloc", "dashcore_hashes/core2", "secp256k1/alloc"]
 
 [package.metadata.docs.rs]
 all-features = true
@@ -52,7 +52,6 @@ internals = { path = "../internals", package = "dashcore-private" }
 bech32 = { version = "0.9.1", default-features = false }
 dashcore_hashes = { path = "../hashes", default-features = false }
 secp256k1 = { default-features = false, features = ["hashes"], version= "0.30.0" }
-dash-network = { path = "../dash-network", default-features = false }
 core2 = { version = "0.4.0", optional = true, features = ["alloc"], default-features = false }
 rustversion = { version="1.0.20"}
 serde = { version = "1.0.219", default-features = false, features = [ "derive", "alloc" ], optional = true }
