+++ dash/src/address.rs
@@ -59,10 +59,10 @@ use crate::blockdata::script::{
 use crate::crypto::key::{PublicKey, TapTweak, TweakedPublicKey, UntweakedPublicKey};
 use crate::error::ParseIntError;
 use crate::hash_types::{PubkeyHash, ScriptHash};
+use crate::network::constants::Network;
 use crate::prelude::*;
 use crate::taproot::TapNodeHash;
 use bech32;
-use dash_network::Network;
 use hashes::{Hash, HashEngine, sha256};
 use internals::write_err;
 use secp256k1::{Secp256k1, Verification, XOnlyPublicKey};
@@ -994,18 +994,15 @@ impl<V: NetworkValidation> Address<V> {
         let p2pkh_prefix = match self.network() {
             Network::Dash => PUBKEY_ADDRESS_PREFIX_MAIN,
             Network::Testnet | Network::Devnet | Network::Regtest => PUBKEY_ADDRESS_PREFIX_TEST,
-            other => unreachable!("Unknown network {other:?} – add explicit prefix"),
         };
         let p2sh_prefix = match self.network() {
             Network::Dash => SCRIPT_ADDRESS_PREFIX_MAIN,
             Network::Testnet | Network::Devnet | Network::Regtest => SCRIPT_ADDRESS_PREFIX_TEST,
-            other => unreachable!("Unknown network {other:?} – add explicit prefix"),
         };
         let bech32_hrp = match self.network() {
             Network::Dash => "ds",
             Network::Testnet | Network::Devnet => "tb",
             Network::Regtest => "dsrt",
-            other => unreachable!("Unknown network {other:?} – add explicit prefix"),
         };
         let encoding = AddressEncoding {
             payload: self.payload(),
@@ -1253,7 +1250,6 @@ impl Address<NetworkUnchecked> {
             (Network::Dash, _) | (_, Network::Dash) => false,
             (Network::Regtest, _) | (_, Network::Regtest) if !is_legacy => false,
             (Network::Testnet, _) | (Network::Regtest, _) | (Network::Devnet, _) => true,
-            _ => false,
         }
     }
 
@@ -1466,8 +1462,8 @@ mod tests {
     use secp256k1::XOnlyPublicKey;
 
     use super::*;
+    use crate::Network::{Dash, Testnet};
     use crate::crypto::key::PublicKey;
-    use dash_network::Network::{Dash, Testnet};
 
     fn roundtrips(addr: &Address) {
         assert_eq!(
