+++ dash/src/crypto/key.rs
@@ -30,10 +30,10 @@ use internals::write_err;
 pub use secp256k1::{self, Keypair, Parity, Secp256k1, Verification, XOnlyPublicKey, constants};
 
 use crate::hash_types::{PubkeyHash, WPubkeyHash};
+use crate::network::constants::Network;
 use crate::prelude::*;
 use crate::taproot::{TapNodeHash, TapTweakHash};
 use crate::{base58, io};
-use dash_network::Network;
 
 /// A key-related error.
 #[derive(Clone, PartialEq, Eq, PartialOrd, Ord, Hash, Debug)]
@@ -407,7 +407,6 @@ impl PrivateKey {
         ret[0] = match self.network {
             Network::Dash => 204,
             Network::Testnet | Network::Devnet | Network::Regtest => 239,
-            _ => 239,
         };
         ret[1..33].copy_from_slice(&self.inner[..]);
         let privkey = if self.compressed {
@@ -820,9 +819,9 @@ mod tests {
     use secp256k1::Secp256k1;
 
     use super::*;
+    use crate::Network::{Dash, Testnet};
     use crate::address::Address;
     use crate::io;
-    use dash_network::Network::{Dash, Testnet};
 
     #[test]
     fn test_key_derivation() {
