+++ dash/src/lib.rs
@@ -157,11 +157,11 @@ pub use crate::hash_types::{
     TxMerkleNode, Txid, WPubkeyHash, WScriptHash, Wtxid,
 };
 pub use crate::merkle_tree::MerkleBlock;
+pub use crate::network::constants::Network;
 pub use crate::pow::{CompactTarget, Target, Work};
 pub use crate::transaction::outpoint::OutPoint;
 pub use crate::transaction::txin::TxIn;
 pub use crate::transaction::txout::TxOut;
-pub use dash_network::Network;
 
 #[cfg(not(feature = "std"))]
 mod io_extras {
