+++ dash/src/sml/llmq_type/network.rs
@@ -1,5 +1,5 @@
+use crate::Network;
 use crate::sml::llmq_type::{DKGWindow, LLMQType};
-use dash_network::Network;
 use std::collections::BTreeMap;
 
 /// Extension trait for Network to add LLMQ-specific methods
@@ -20,7 +20,6 @@ impl NetworkLLMQExt for Network {
             Network::Testnet => LLMQType::Llmqtype50_60,
             Network::Devnet => LLMQType::LlmqtypeDevnet,
             Network::Regtest => LLMQType::LlmqtypeTestInstantSend,
-            other => unreachable!("Unsupported network variant {other:?}"),
         }
     }
 
@@ -30,7 +29,6 @@ impl NetworkLLMQExt for Network {
             Network::Testnet => LLMQType::Llmqtype60_75,
             Network::Devnet => LLMQType::LlmqtypeDevnetDIP0024,
             Network::Regtest => LLMQType::LlmqtypeTestDIP0024,
-            other => unreachable!("Unsupported network variant {other:?}"),
         }
     }
 
@@ -40,7 +38,6 @@ impl NetworkLLMQExt for Network {
             Network::Testnet => LLMQType::Llmqtype50_60,
             Network::Devnet => LLMQType::LlmqtypeDevnet,
             Network::Regtest => LLMQType::LlmqtypeTest,
-            other => unreachable!("Unsupported network variant {other:?}"),
         }
     }
 
@@ -50,7 +47,6 @@ impl NetworkLLMQExt for Network {
             Network::Testnet => LLMQType::Llmqtype25_67,
             Network::Devnet => LLMQType::LlmqtypeDevnet,
             Network::Regtest => LLMQType::LlmqtypeTest,
-            other => unreachable!("Unsupported network variant {other:?}"),
         }
     }
 
@@ -80,7 +76,6 @@ impl NetworkLLMQExt for Network {
                 LLMQType::LlmqtypeTestDIP0024,
                 LLMQType::LlmqtypeTestInstantSend,
             ],
-            other => unreachable!("Unsupported network variant {other:?}"),
         }
     }
