+++ key-wallet-ffi/Cargo.toml
@@ -23,7 +23,6 @@ bls = ["dashcore/bls", "key-wallet/bls"]
 key-wallet = { path = "../key-wallet", default-features = false, features = ["std"] }
 key-wallet-manager = { path = "../key-wallet-manager", features = ["std"] }
 dashcore = { path = "../dash", features = ["std"] }
-dash-network = { path = "../dash-network" }
 secp256k1 = { version = "0.30.0", features = ["global-context"] }
 tokio = { version = "1.32", features = ["rt-multi-thread", "sync"] }
 libc = "0.2"
