+++ key-wallet/Cargo.toml
@@ -10,9 +10,9 @@ license = "CC0-1.0"
 
 [features]
 default = ["std"]
-std = ["dashcore_hashes/std", "secp256k1/std", "bip39/std", "getrandom", "dash-network/std", "rand"]
-serde = ["dep:serde", "dep:serde_json", "dashcore_hashes/serde", "secp256k1/serde", "dash-network/serde", "dashcore/serde"]
-bincode = ["serde", "dep:bincode", "dep:bincode_derive", "dash-network/bincode", "dashcore_hashes/bincode", "dashcore/bincode"]
+std = ["dashcore_hashes/std", "secp256k1/std", "bip39/std", "getrandom", "rand"]
+serde = ["dep:serde", "dep:serde_json", "dashcore_hashes/serde", "secp256k1/serde", "dashcore/serde"]
+bincode = ["serde", "dep:bincode", "dep:bincode_derive", "dashcore_hashes/bincode", "dashcore/bincode"]
 bip38 = ["scrypt", "aes", "bs58", "rand"]
 eddsa = ["dashcore/eddsa"]
 bls = ["dashcore/bls"]
@@ -28,7 +28,6 @@ serde = { version = "1.0", default-features = false, features = ["derive"], opti
 base58ck = { version = "0.1.0", default-features = false }
 bitflags = { version = "2.6", default-features = false }
 getrandom = { version = "0.2", optional = true }
-dash-network = { path = "../dash-network", default-features = false }
 # BIP38 dependencies (optional)
 scrypt = { version = "0.11", default-features = false, optional = true }
 aes = { version = "0.8", default-features = false, optional = true }
