+++ key-wallet/src/derivation_slip10.rs
@@ -13,8 +13,8 @@ use core::fmt;
 use std::error;
 
 use alloc::{string::String, vec::Vec};
-use dash_network::Network;
 pub use dashcore::ed25519_dalek::{SigningKey, VerifyingKey};
+use dashcore::Network;
 use dashcore_hashes::{sha512, Hash, HashEngine, Hmac, HmacEngine};
 #[cfg(feature = "serde")]
 use serde;
