+++ key-wallet/src/lib.rs
@@ -57,7 +57,7 @@ pub use account::{Account, AccountCollection, AccountType};
 pub use bip32::{ChildNumber, DerivationPath, ExtendedPrivKey, ExtendedPubKey};
 #[cfg(feature = "bip38")]
 pub use bip38::{encrypt_private_key, generate_intermediate_code, Bip38EncryptedKey, Bip38Mode};
-pub use dash_network::Network;
+pub use dashcore::Network;
 pub use dashcore::{Address, AddressType};
 pub use derivation::{DerivationPathBuilder, DerivationStrategy, KeyDerivation};
 pub use dip9::{DerivationPathReference, DerivationPathType};
